GH-769: Polishing properties - moved to separate package
