+++ master/src/main/java/com/axelixlabs/axelix/master/autoconfiguration/auth/properties/StaticAdminCredentialsProperties.java
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package com.axelixlabs.axelix.master.autoconfiguration.auth;
+package com.axelixlabs.axelix.master.autoconfiguration.auth.properties;
 
 /**
  * Configuration of the static-admin.
