+++ master/src/main/java/com/axelixlabs/axelix/master/service/auth/provider/StaticAdminUserProvider.java
@@ -23,7 +23,7 @@ import java.util.Set;
 import com.axelixlabs.axelix.common.auth.core.DefaultRole;
 import com.axelixlabs.axelix.common.auth.core.DefaultUser;
 import com.axelixlabs.axelix.common.auth.core.User;
-import com.axelixlabs.axelix.master.autoconfiguration.auth.StaticAdminCredentialsProperties;
+import com.axelixlabs.axelix.master.autoconfiguration.auth.properties.StaticAdminCredentialsProperties;
 import com.axelixlabs.axelix.master.exception.auth.UserNotFoundException;
 import com.axelixlabs.axelix.master.service.auth.UserLoginService;
